refactor: relocate the files in candidates
