+++ app/src/main/java/com/osfans/trime/ime/candidates/unrolled/PagingCandidateViewAdapter.kt
@@ -1,8 +1,9 @@
-// SPDX-FileCopyrightText: 2015 - 2024 Rime community
-//
-// SPDX-License-Identifier: GPL-3.0-or-later
+/*
+ * SPDX-FileCopyrightText: 2015 - 2024 Rime community
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
 
-package com.osfans.trime.ime.candidates.adapter
+package com.osfans.trime.ime.candidates.unrolled
 
 import android.view.ViewGroup
 import androidx.paging.PagingDataAdapter
