+++ app/src/main/java/com/osfans/trime/ime/candidates/unrolled/window/FlexboxUnrolledCandidateWindow.kt
@@ -16,8 +16,8 @@ import com.osfans.trime.daemon.RimeSession
 import com.osfans.trime.data.theme.Theme
 import com.osfans.trime.ime.bar.QuickBar
 import com.osfans.trime.ime.candidates.CandidateViewHolder
-import com.osfans.trime.ime.candidates.CompactCandidateModule
-import com.osfans.trime.ime.candidates.adapter.PagingCandidateViewAdapter
+import com.osfans.trime.ime.candidates.compact.CompactCandidateModule
+import com.osfans.trime.ime.candidates.unrolled.PagingCandidateViewAdapter
 import com.osfans.trime.ime.candidates.unrolled.UnrolledCandidateLayout
 import com.osfans.trime.ime.candidates.unrolled.decoration.FlexboxHorizontalDecoration
 import com.osfans.trime.ime.core.TrimeInputMethodService
