+++ app/src/main/java/com/osfans/trime/ime/dependency/InputComponent.kt
@@ -10,7 +10,7 @@ import com.osfans.trime.data.theme.Theme
 import com.osfans.trime.ime.bar.QuickBar
 import com.osfans.trime.ime.broadcast.EnterKeyLabelModule
 import com.osfans.trime.ime.broadcast.InputBroadcaster
-import com.osfans.trime.ime.candidates.CompactCandidateModule
+import com.osfans.trime.ime.candidates.compact.CompactCandidateModule
 import com.osfans.trime.ime.composition.CompositionPopupWindow
 import com.osfans.trime.ime.core.InputView
 import com.osfans.trime.ime.core.TrimeInputMethodService
